requirements.txt
# API and Web Framework
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Data Processing and Analysis
pandas==2.1.2
numpy==1.26.1
scipy==1.11.3

# Visualization
matplotlib==3.8.1
seaborn==0.13.0
plotly==5.18.0

# Frontend
streamlit==1.28.1
pillow==10.1.0

# HTTP Client
requests==2.31.0

# Utilities
python-multipart==0.0.6
email-validator==2.1.0

# Testing
pytest==7.4.3 